requirements.txt
python-binance==1.0.35
requests==2.32.5
pymysql==1.1.2
websocket-client==1.9.0
python-dotenv==1.2.1
